Instant discounts: review discount percentage calculation + cleanup

+++ src/MEGASync/control/Utilities.cpp
@@ -24,6 +24,7 @@
 #include <QScreen>
 #include <QTextStream>
 
+#include <cmath>
 #include <iostream>
 
 #ifndef WIN32
@@ -2233,13 +2234,13 @@ QString Utilities::toPrice(double value, const QString& currencySymbol, bool sho
     return price;
 }
 
-int Utilities::softCeil(double value)
+double Utilities::softCeil(double value)
 {
     // softCeil(number) {
     //     'use strict';
     //     return Math.ceil(Math.round(number * 10) / 10);
     // }
-    return static_cast<int>(std::ceil(std::round(value * 10.) / 10.));
+    return std::ceil(std::round(value * 10.) / 10.);
 }
 
 void MegaListenerFuncExecuter::setExecuteInAppThread(bool executeInAppThread)
