+++ src/MEGASync/control/Utilities.h
@@ -632,7 +632,7 @@ public:
 
     static QString decodeUnicodeEscapes(const QString& input);
     static QString toPrice(double value, const QString& currencySymbol, bool showRemark = false);
-    static int softCeil(double value);
+    static double softCeil(double value);
 
     // Human-friendly list of forbidden chars for New Remote Folder
     static const QLatin1String FORBIDDEN_CHARS;
